Split hardware.py into a few files
